match 'int16_t' which is passed back by `parser.value_int()`, and take advantage of `parser.value_byte()`

+++ Marlin/src/feature/bedlevel/ubl/ubl.h
@@ -47,7 +47,7 @@ struct mesh_index_pair;
 
 typedef struct {
   bool      C_seen;
-  int8_t    KLS_storage_slot,
+  int16_t   KLS_storage_slot,
             T_map_type;
   uint8_t   R_repetition,
             V_verbosity,
